+++ word_pasting.py
@@ -1,16 +1,10 @@
 from collections import defaultdict, Counter
 
-import data
 import random
-import os
-import numpy as np
 import json
 import sys
 
-
-def create_random_word(word_length: int, char_repertoire: str, weights: list) -> str:
-    assert len(char_repertoire) == len(weights)
-    return ''.join(random.choices(char_repertoire, weights=weights, k=word_length))
+from compression_entropy import read_selected_verses, run_mismatcher, get_entropy, to_file, create_random_word
 
 
 def mask_word_structure(tokenized: list, char_str: str, char_weights: list) -> list:
@@ -56,41 +50,8 @@ def replace_words(verse_tokens: list) -> list:
     return verse_chars
 
 
-# TODO: remove files after running
-def to_file(text: str, base_filename: str, appendix: str) -> str:
-    """
-    Save a text to a file
-    :param text: the text to be saved
-    :param base_filename: base filename to be used
-    :param appendix: appendix to be added to this filename
-    :return: the new filename created
-    """
-    dot_parts = base_filename.split('.')
-    extension = dot_parts[-1]
-    prefix = '.'.join(dot_parts[:-1])
-    new_filename = prefix + '_' + appendix + '.' + extension
-    with open(new_filename, 'w') as f:
-        f.write(text)
-    return new_filename
-
-
-def run_mismatcher(preprocessed_filename: str, remove_file: bool, executable_path: str) -> list:
-    mismatcher_filename = preprocessed_filename + '_mismatcher'
-    os.system(f"""java -Xmx3500M -jar \
-                {executable_path} {preprocessed_filename} {mismatcher_filename}""")
-    with open(mismatcher_filename, 'r') as f:
-        lines = f.readlines()
-    if remove_file:
-        os.remove(mismatcher_filename)
-    return parse_mismatcher_lines(lines)
-
-
 def parse_mismatcher_lines(lines: list) -> list:
     return [int(line.split('\t')[-1].strip()) for line in lines if line != '\n']
-
-
-def get_entropy(mismatches: list) -> float:
-    return 1 / (sum([el / np.log2(i + 2) for i, el in enumerate(mismatches[1:])]) / len(mismatches))
 
 
 def get_text_length(sequences: list) -> int:
@@ -207,23 +168,6 @@ def get_entropies_per_word(sample_verses: list,
 
 def get_char_distribution(text: str) -> dict:
     return Counter(text)
-
-
-def read_selected_verses(filename: str,
-                         lowercase: bool,
-                         chosen_books: list,
-                         truncate_books: bool) -> tuple:
-    # Read the complete bible
-    bible = data.parse_pbc_bible(filename)
-    # Tokenize by splitting on spaces
-    tokenized = bible.tokenize(remove_punctuation=False, lowercase=lowercase)
-    # Obtain the repertoire of symbols
-    char_counter = get_char_distribution(''.join([el for lis in tokenized.verse_tokens.values() for el in lis]))
-    # Split by book
-    _, _, book_verses, _, _ = data.join_by_toc(tokenized.verse_tokens)
-    # Select the books we are interested in
-    selected_book_verses = select_samples(book_verses, chosen_books, truncate_books)
-    return selected_book_verses, char_counter
 
 
 def join_words(verse: list, locations: list) -> list:
